pyproject.toml
[tool.poetry]
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"
fastapi-pagination = "^0.12.34"
kafka-python-ng = "^2.2.3"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
pytest = "^8.3.3"
pylint = "^3.3.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
faker = "^33.1.0"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# black
[tool.black]
line-length = 79

# isort
[tool.isort]
profile = "black"
known_local_folder=["api"]
line_length = 79
lines_after_imports = 2


[tool.pylint]
ignore = [".git", "__pycache__", "migrations", ".venv"]
max-line-length = 79
disable = ["C0114", "C0115", "C0116", "W1514", "W3101", "R0903", "C0103"]